+++ CHANGELOG.md
@@ -11,8 +11,15 @@ created.
 ## Upcoming
 
 ### Added
+- ([#385](https://github.com/microsoft/InnerEye-DeepLearning/pull/385)) Add the ability to train a model on multiple
+nodes in AzureML. Example: Add `--num_nodes=2` to the commandline arguments to train on 2 nodes.
 
 ### Changed
+- ([#385](https://github.com/microsoft/InnerEye-DeepLearning/pull/385)) Starting an AzureML run now uses the
+`ScriptRunConfig` object, rather than the deprecated `Estimator` object.
+- ([#385](https://github.com/microsoft/InnerEye-DeepLearning/pull/385)) When registering a model, the name of the 
+Python execution environment is added as a tag. This tag is read when running inference, and the execution environment
+is re-used.
 
 ### Fixed